requirements.txt
appdirs==1.4.4
buildozer==1.2.0
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
docutils==0.16
filelock==3.0.12
idna==2.10
importlib-metadata==3.4.0
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
pexpect==4.8.0
Pillow==8.1.0
ptyprocess==0.7.0
Pygments==2.7.4
pypiwin32==223
python-dotenv==0.15.0
pywin32==300
requests==2.25.1
sh==1.14.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.0
zipp==3.4.0
